move add goal stuff
